Moved around some tests for organizational purposes
