+++ osquery/core/test_util.h
@@ -22,6 +22,10 @@
 #include <osquery/database.h>
 #include <osquery/filesystem.h>
 
+// These codes are specifically for process abstraction testing
+#define EXTENSION_SUCCESS_CODE  0x45
+#define WORKER_SUCCESS_CODE     0x57
+
 namespace pt = boost::property_tree;
 
 namespace osquery {
