+++ osquery/core/tests/process_tests.cpp
@@ -11,11 +11,8 @@
 #include <osquery/core.h>
 #include <gtest/gtest.h>
 
-#ifdef WIN32
-#include <Shlwapi.h>
-#endif
-
 #include "osquery/core/process.h"
+#include "osquery/core/test_util.h"
 
 extern char *self_exec_path;
 
@@ -56,98 +53,9 @@ bool getProcessExitCode(osquery::PlatformProcess& process, int& exitCode)
   return false;
 }
 
-#define EXTENSION_SUCCESS_CODE  0x45
-#define WORKER_SUCCESS_CODE     0x57
-
-#define OSQUERY_TESTS_MODULENAME  "osquery_tests.exe"
-
-const char *expected_worker_args[] = {
-  "worker-test"
-};
-const char *expected_extension_args[] = {
-  "osquery extension: extension-test",
-  "--socket",
-  "socket-name",
-  "--timeout",
-  "100",
-  "--interval",
-  "5",
-  "--verbose"
-};
-
-bool compareArguments(char *result[], 
-                      unsigned int result_nelms,
-                      const char *expected[], 
-                      unsigned int expected_nelms) {
-  if (result_nelms != expected_nelms) {
-    return false;
-  }
-
-  for (int i = 0; i < expected_nelms; i++) {
-    if (strlen(result[i]) != strlen(expected[i])) {
-      return false;
-    }
-
-    if (strncmp(result[i], expected[i], strlen(expected[i])) != 0) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-int workerMain(int argc, char *argv[]) {
-  if (!compareArguments(argv,
-                        argc,
-                        expected_worker_args,
-                        sizeof(expected_worker_args) / sizeof(const char *))) {
-    return -1;
-  }
-
-  osquery::PlatformProcess process = osquery::getLauncherProcess();
-  if (!process.isValid()) {
-    return -2;
-  }
-
-#ifdef WIN32
-  CHAR buffer[1024] = { 0 };
-  DWORD size = 1024;
-  if (!QueryFullProcessImageNameA(process.nativeHandle(),
-                                  0,
-                                  buffer,
-                                  &size)) {
-    return -3;
-  }
-  PathStripPathA(buffer);
-
-  if (strlen(buffer) != strlen(OSQUERY_TESTS_MODULENAME)) {
-    return -4;
-  }
-
-  if (strncmp(buffer, OSQUERY_TESTS_MODULENAME, strlen(buffer)) != 0) {
-    return -5;
-  }
-#else
-  if (process.nativeHandle() != getppid()) {
-    return -3;
-  }
-#endif
-  return WORKER_SUCCESS_CODE;
-}
-
-int extensionMain(int argc, char *argv[]) {
-  if (!compareArguments(argv, 
-                        argc, 
-                        expected_extension_args, 
-                        sizeof(expected_extension_args) / sizeof(const char *))) {
-    return -1;
-  }
-  return EXTENSION_SUCCESS_CODE;
-}
-
 namespace osquery {
 
-class ProcessTests : public testing::Test {};
+class ProcessTests : public testing::Test { };
 
 TEST_F(ProcessTests, test_constructor) {
   auto p = PlatformProcess(kInvalidPid);
